+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # v0.8.0-dev
+* Enhancements
+  * Can now support catch blocks in try expressions
 
 # v0.7.0
 * Enhancements